% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{create_application_version}
\alias{create_application_version}
\title{Creates an application version}
\usage{
create_application_version(ApplicationId, SemanticVersion,
  SourceCodeUrl = NULL, TemplateBody = NULL, TemplateUrl = NULL)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{SemanticVersion}{[required] The semantic version of the new version.}

\item{SourceCodeUrl}{A link to a public repository for the source code of your application.}

\item{TemplateBody}{The raw packaged AWS SAM template of your application.}

\item{TemplateUrl}{A link to the packaged AWS SAM template of your application.}
}
\description{
Creates an application version.
}
\section{Accepted Parameters}{
\preformatted{create_application_version(
  ApplicationId = "string",
  SemanticVersion = "string",
  SourceCodeUrl = "string",
  TemplateBody = "string",
  TemplateUrl = "string"
)
}
}

