% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{delete_application}
\alias{delete_application}
\title{Deletes the specified application}
\usage{
delete_application(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}
}
\description{
Deletes the specified application.
}
\section{Accepted Parameters}{
\preformatted{delete_application(
  ApplicationId = "string"
)
}
}

