% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{get_application}
\alias{get_application}
\title{Gets the specified application}
\usage{
get_application(ApplicationId, SemanticVersion = NULL)
}
\arguments{
\item{ApplicationId}{The Amazon Resource Name (ARN) of the application.}

\item{SemanticVersion}{The semantic version of the application to get.}
}
\description{
Gets the specified application.
}
\examples{

}
