% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{get_application_policy}
\alias{get_application_policy}
\title{Retrieves the policy for the application}
\usage{
get_application_policy(ApplicationId)
}
\arguments{
\item{ApplicationId}{The Amazon Resource Name (ARN) of the application.}
}
\description{
Retrieves the policy for the application.
}
\examples{

}
