% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{get_cloud_formation_template}
\alias{get_cloud_formation_template}
\title{Gets the specified AWS CloudFormation template}
\usage{
get_cloud_formation_template(
  ApplicationId = "string",
  TemplateId = "string"
)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{TemplateId}{[required] The UUID returned by CreateCloudFormationTemplate.

Pattern: [0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}}
}
\description{
Gets the specified AWS CloudFormation template.
}
\examples{

}
