% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{list_application_dependencies}
\alias{list_application_dependencies}
\title{Retrieves the list of applications nested in the containing application}
\usage{
list_application_dependencies(ApplicationId, MaxItems = NULL,
  NextToken = NULL, SemanticVersion = NULL)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}

\item{SemanticVersion}{The semantic version of the application to get.}
}
\description{
Retrieves the list of applications nested in the containing application.
}
\section{Accepted Parameters}{
\preformatted{list_application_dependencies(
  ApplicationId = "string",
  MaxItems = 123,
  NextToken = "string",
  SemanticVersion = "string"
)
}
}

