% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{list_application_versions}
\alias{list_application_versions}
\title{Lists versions for the specified application}
\usage{
list_application_versions(ApplicationId, MaxItems = NULL,
  NextToken = NULL)
}
\arguments{
\item{ApplicationId}{[required] The Amazon Resource Name (ARN) of the application.}

\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}
}
\description{
Lists versions for the specified application.
}
\section{Accepted Parameters}{
\preformatted{list_application_versions(
  ApplicationId = "string",
  MaxItems = 123,
  NextToken = "string"
)
}
}

