% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{list_applications}
\alias{list_applications}
\title{Lists applications owned by the requester}
\usage{
list_applications(MaxItems = NULL, NextToken = NULL)
}
\arguments{
\item{MaxItems}{The total number of items to return.}

\item{NextToken}{A token to specify where to start paginating.}
}
\description{
Lists applications owned by the requester.
}
\section{Accepted Parameters}{
\preformatted{list_applications(
  MaxItems = 123,
  NextToken = "string"
)
}
}

