% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.serverlessapplicationrepository_package.R
\docType{package}
\name{paws.serverlessapplicationrepository-package}
\alias{paws.serverlessapplicationrepository}
\alias{paws.serverlessapplicationrepository-package}
\title{paws.serverlessapplicationrepository: AWSServerlessApplicationRepository}
\description{
The AWS Serverless Application Repository makes it
    easy for developers and enterprises to quickly find and deploy
    serverless applications in the AWS Cloud. For more information about
    serverless applications, see Serverless Computing and Applications on
    the AWS website.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
