% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/paws.serverlessapplicationrepository_operations.R
\name{put_application_policy}
\alias{put_application_policy}
\title{Sets the permission policy for an application}
\usage{
put_application_policy(ApplicationId, Statements)
}
\arguments{
\item{ApplicationId}{The Amazon Resource Name (ARN) of the application.}

\item{Statements}{An array of policy statements applied to the application.}
}
\description{
Sets the permission policy for an application. See \href{https://docs.aws.amazon.com/serverlessrepo/latest/devguide/access-control-resource-based.html#application-permissions}{Application Permissions} for the list of supported actions that can be used with this operation.
}
\examples{

}
